pyproject.toml
[project]
name = "pycmds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "python-magic>=0.4.27",
]
